security/requirements.txt
pytest>=7.3.1
pyyaml>=6.0.0
selenium>=4.15.0
webdriver-manager>=4.0.0
requests>=2.31.0
loguru>=0.7.0
